+++ src/codecs/webp/extended.rs
@@ -1,5 +1,5 @@
 use std::convert::TryInto;
-use std::io::{self, Error, Read};
+use std::io::{self, Cursor, Error, Read};
 use std::{error, fmt};
 
 use super::decoder::{read_chunk, DecoderError::ChunkHeaderInvalid, WebPRiffChunk};
@@ -7,7 +7,7 @@ use super::lossless::{LosslessDecoder, LosslessFrame};
 use super::vp8::{Frame as VP8Frame, Vp8Decoder};
 use crate::error::DecodingError;
 use crate::image::ImageFormat;
-use crate::{color, Delay, Frame, Frames, ImageError, ImageResult, Rgba, RgbaImage};
+use crate::{color, Delay, Frame, Frames, ImageError, ImageResult, Rgb, RgbImage, Rgba, RgbaImage};
 use byteorder::{LittleEndian, ReadBytesExt};
 
 //all errors that can occur while parsing extended chunks in a WebP file
@@ -164,7 +164,7 @@ impl ExtendedImage {
                         let alpha_chunk =
                             read_alpha_chunk(&mut cursor, info.canvas_width, info.canvas_height)?;
 
-                        let vp8_frame = read_lossy(reader)?;
+                        let vp8_frame = read_lossy_with_chunk(reader)?;
 
                         let img = WebPStatic::from_alpha_lossy(alpha_chunk, vp8_frame)?;
 
@@ -173,7 +173,7 @@ impl ExtendedImage {
                 }
                 WebPRiffChunk::VP8 => {
                     if static_frame.is_none() {
-                        let vp8_frame = read_lossy(&mut cursor)?;
+                        let vp8_frame = read_lossy(cursor)?;
 
                         let img = WebPStatic::from_lossy(vp8_frame)?;
 
@@ -329,7 +329,8 @@ impl ExtendedImage {
 
 #[derive(Debug)]
 enum WebPStatic {
-    Lossy(RgbaImage),
+    LossyWithAlpha(RgbaImage),
+    LossyWithoutAlpha(RgbImage),
     Lossless(LosslessFrame),
 }
 
@@ -375,7 +376,7 @@ impl WebPStatic {
         let image = RgbaImage::from_vec(vp8_frame.width.into(), vp8_frame.height.into(), image_vec)
             .unwrap();
 
-        Ok(WebPStatic::Lossy(image))
+        Ok(WebPStatic::LossyWithAlpha(image))
     }
 
     fn get_predictor(
@@ -441,21 +442,24 @@ impl WebPStatic {
     }
 
     pub(crate) fn from_lossy(vp8_frame: VP8Frame) -> ImageResult<WebPStatic> {
-        let mut image = RgbaImage::from_pixel(
+        let mut image = RgbImage::from_pixel(
             vp8_frame.width.into(),
             vp8_frame.height.into(),
-            Rgba([0, 0, 0, 255]),
+            Rgb([0, 0, 0]),
         );
 
-        vp8_frame.fill_rgba(&mut image);
+        vp8_frame.fill_rgb(&mut image);
 
-        Ok(WebPStatic::Lossy(image))
+        Ok(WebPStatic::LossyWithoutAlpha(image))
     }
 
     pub(crate) fn fill_buf(&self, buf: &mut [u8]) {
         match self {
-            WebPStatic::Lossy(image) => {
-                buf.copy_from_slice(&**image);
+            WebPStatic::LossyWithAlpha(image) => {
+                buf.copy_from_slice(image);
+            }
+            WebPStatic::LossyWithoutAlpha(image) => {
+                buf.copy_from_slice(image);
             }
             WebPStatic::Lossless(lossless) => {
                 lossless.fill_rgba(buf);
@@ -465,7 +469,8 @@ impl WebPStatic {
 
     pub(crate) fn get_buf_size(&self) -> usize {
         match self {
-            WebPStatic::Lossy(lossy) => lossy.len(),
+            WebPStatic::LossyWithAlpha(rgb_image) => rgb_image.len(),
+            WebPStatic::LossyWithoutAlpha(rgba_image) => rgba_image.len(),
             WebPStatic::Lossless(lossless) => lossless.get_buf_size(),
         }
     }
@@ -593,7 +598,7 @@ fn read_3_bytes<R: Read>(reader: &mut R) -> ImageResult<u32> {
     Ok(value)
 }
 
-fn read_lossy<R: Read>(reader: &mut R) -> ImageResult<VP8Frame> {
+fn read_lossy_with_chunk<R: Read>(reader: &mut R) -> ImageResult<VP8Frame> {
     let (cursor, chunk) =
         read_chunk(reader)?.ok_or_else(|| Error::from(io::ErrorKind::UnexpectedEof))?;
 
@@ -601,6 +606,10 @@ fn read_lossy<R: Read>(reader: &mut R) -> ImageResult<VP8Frame> {
         return Err(ChunkHeaderInvalid(chunk.to_fourcc()).into());
     }
 
+    read_lossy(cursor)
+}
+
+fn read_lossy(cursor: Cursor<Vec<u8>>) -> ImageResult<VP8Frame> {
     let mut vp8_decoder = Vp8Decoder::new(cursor);
     let frame = vp8_decoder.decode_frame()?;
 
@@ -630,7 +639,7 @@ fn read_image<R: Read>(reader: &mut R, width: u32, height: u32) -> ImageResult<W
         Some((mut cursor, WebPRiffChunk::ALPH)) => {
             let alpha_chunk = read_alpha_chunk(&mut cursor, width, height)?;
 
-            let vp8_frame = read_lossy(reader)?;
+            let vp8_frame = read_lossy_with_chunk(reader)?;
 
             let img = WebPStatic::from_alpha_lossy(alpha_chunk, vp8_frame)?;