+++ src/imageops/sample.rs
@@ -7,7 +7,7 @@ use std::f32;
 
 use num_traits::{NumCast, ToPrimitive, Zero};
 
-use crate::image::GenericImageView;
+use crate::image::{GenericImage, GenericImageView};
 use crate::traits::{Enlargeable, Pixel, Primitive};
 use crate::utils::clamp;
 use crate::{ImageBuffer, Rgba32FImage};
@@ -110,7 +110,7 @@ struct FloatNearest(f32);
 // to_i64, to_u64, and to_f64 implicitly affect all other lower conversions.
 // Note that to_f64 by default calls to_i64 and thus needs to be overridden.
 impl ToPrimitive for FloatNearest {
-    // to_{i,u}64 is required, to_{i,u}{8,16} are usefull.
+    // to_{i,u}64 is required, to_{i,u}{8,16} are useful.
     // If a usecase for full 32 bits is found its trivial to add
     fn to_i8(&self) -> Option<i8> {
         self.0.round().to_i8()
@@ -760,6 +760,13 @@ where
     I::Pixel: 'static,
     <I::Pixel as Pixel>::Subpixel: 'static,
 {
+    // check if the new dimensions are the same as the old. if they are, make a copy instead of resampling
+    if (nwidth, nheight) == image.dimensions() {
+        let mut tmp = ImageBuffer::new(image.width(), image.height());
+        tmp.copy_from(image, 0, 0).unwrap();
+        return tmp;
+    }
+
     let mut method = match filter {
         FilterType::Nearest => Filter {
             kernel: Box::new(box_kernel),
@@ -860,7 +867,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::{resize, FilterType};
-    use crate::{ImageBuffer, RgbImage};
+    use crate::{GenericImageView, ImageBuffer, RgbImage};
     #[cfg(feature = "benchmarks")]
     use test;
 
@@ -873,6 +880,29 @@ mod tests {
             test::black_box(resize(&img, 200, 200, FilterType::Nearest));
         });
         b.bytes = 800 * 800 * 3 + 200 * 200 * 3;
+    }
+
+    #[test]
+    #[cfg(feature = "png")]
+    fn test_resize_same_size() {
+        use std::path::Path;
+        let img = crate::open(&Path::new("./examples/fractal.png")).unwrap();
+        let resize = img.resize(img.width(), img.height(), FilterType::Triangle);
+        assert!(img.pixels().eq(resize.pixels()))
+    }
+
+    #[bench]
+    #[cfg(all(feature = "benchmarks", feature = "tiff"))]
+    fn bench_resize_same_size(b: &mut test::Bencher) {
+        let path = concat!(
+            env!("CARGO_MANIFEST_DIR"),
+            "/tests/images/tiff/testsuite/mandrill.tiff"
+        );
+        let image = crate::open(path).unwrap();
+        b.iter(|| {
+            test::black_box(image.resize(image.width(), image.height(), FilterType::CatmullRom));
+        });
+        b.bytes = (image.width() * image.height() * 3) as u64;
     }
 
     #[test]