+++ src/lib.rs
@@ -187,7 +187,7 @@ pub mod flat;
 /// | BMP    | Yes | Rgb8, Rgba8, Gray8, GrayA8 |
 /// | ICO    | Yes | Yes |
 /// | TIFF   | Baseline(no fax support) + LZW + PackBits | Rgb8, Rgba8, Gray8 |
-/// | WebP   | Yes | No |
+/// | WebP   | Yes | Rgb8, Rgba8 |
 /// | AVIF   | Only 8-bit | Lossy |
 /// | PNM    | PBM, PGM, PPM, standard PAM | Yes |
 /// | DDS    | DXT1, DXT3, DXT5 | No |
@@ -236,7 +236,7 @@ pub mod codecs {
     pub mod ico;
     #[cfg(feature = "jpeg")]
     pub mod jpeg;
-    #[cfg(feature = "openexr")]
+    #[cfg(feature = "exr")]
     pub mod openexr;
     #[cfg(feature = "png")]
     pub mod png;
@@ -246,7 +246,7 @@ pub mod codecs {
     pub mod tga;
     #[cfg(feature = "tiff")]
     pub mod tiff;
-    #[cfg(feature = "webp")]
+    #[cfg(any(feature = "webp", feature = "webp-encoder"))]
     pub mod webp;
 }
 
@@ -270,6 +270,7 @@ mod utils;
 // Copyright (c) 2018 Guillaume Gomez
 macro_rules! insert_as_doc {
     { $content:expr } => {
+        #[allow(unused_doc_comments)]
         #[doc = $content] extern { }
     }
 }